Correct order date display and order supplements display

+++ app/src/main/java/ooo/cron/delivery/data/network/models/OrderHistoryDetailAdditive.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class OrderHistoryDetailAdditive(
     @SerializedName("Id") val id: String,
     @SerializedName("Name") val name: String,
-    @SerializedName("Cost") val cost: Double
+    @SerializedName("Cost") val cost: Int
 )
